pyproject.toml
[project]
name = "deeplx-tr"
version = "0.1.0a1"
description = " TODO "
authors = [
    { name = "ffreemt" }
]
license = "MIT"
repository = "https://github.com/ffreemt/deeplx-tr"

dependencies = [
    "loguru>=0.7.2",
    "rich>=13.7.1",
    "typer>=0.12.1",
    "httpx>=0.27.0",
    "pyperclip>=1.8.2",
    "set_loglevel>=0.1.2",
    "nest-asyncio>=1.6.0",
    "diskcache>=5.6.3",
    "loadtext>=0.0.3",
    "pyquery>=2.0.0",
    "openai>=1.35.3",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "langchain-openai>=0.1.14",
    "langchain>=0.2.6",
    "taipy>=3.1.1",
    "python-magic>=0.4.27",
    'python-magic-bin>=0.4.14; sys_platform == "win32"',
    "lmtr-agents>=0.1.0a2",
    "streamlit>=1.38.0",
    "python-box>=7.2.0",
    "ycecream>=1.3.17",
    "sentry-sdk[fastapi]>=2.17.0",
    "marimo>=0.10.9",
    "unsync>=1.4.0",
]
readme = "README.md"
requires-python = ">= 3.10"

[project.scripts]
"deeplx-tr" = "deeplx_tr.__main__:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.2.2",
    "ipython>=8.12.3",
    "ruff>=0.3.5",
    "pytest-asyncio>=0.23.7",
    "ycecream>=1.3.17",
    "pylint>=3.2.4",
    "pytest-timeout>=2.3.1",
    "python-run-cmd>=0.1.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/deeplx_tr"]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"
asyncio_mode = "auto"
minversion = "6.0"
# addopts = "-ra -q --doctest-modules -s"
addopts = "-ra -q --doctest-modules -s -p no:warnings"

log_cli = true
testpaths = [
    "tests",
    # "integration",
]

[tool.uv]
dev-dependencies = [
    "pytest-asyncio>=0.24.0",
    "pytest-timeout>=2.3.1",
    "pytest>=8.3.4",
    "python-run-cmd>=0.1.1",
]
